+++ src/components/pages/main_page/display/ai/SargaAITranslate.svelte
