+++ src/components/pages/main_page/display/ai/translation_prompts.yaml
