+++ src/components/pages/main_page/display/bulk/BulkEdit.svelte
