+++ src/components/pages/main_page/display/bulk/trans_bulk_funcs.ts
